Modifies 0-print_listint.c

+++ 0x13-more_singly_linked_lists/0-print_listint.c
@@ -1,14 +1,14 @@
 #include "lists.h"
 
 /**
- * print_listint - Prints all the elements of a listint_t list
+ * print_listint - Prints all the elements of the list
  * @h: A pointer to the start of the list
  *
  * Return: the number of nodes
 */
 size_t print_listint(const listint_t *h)
 {
-	size_t nodes = 0;
+	unsigned int nodes = 0;
 
 	while (h != NULL)
 	{
